pyproject.toml
# SPDX-License-Identifier: GPL-3.0-or-later
# SPDX-FileNotice: Part of the Cubinets addon.


[project]
version = '0.0.0' # No need to update
name = 'Addon'    # these two properties.

requires-python = '>=3.11'

dependencies = [
    'freecad-stubs>=1.0.21',
    'pyside6>=6.9.2'
]
